move label types to new file
