+++ packages/core/src/components/Label/Label.tsx
@@ -7,38 +7,12 @@ import { backwardCompatibilityForProperties } from "../../helpers/backwardCompat
 import Text from "../Text/Text";
 import Leg from "./Leg";
 import { LabelColor, LabelKind } from "./LabelConstants";
-import { VibeComponent, VibeComponentProps, withStaticProps } from "../../types";
+import { VibeComponent, withStaticProps } from "../../types";
 import useClickableProps from "../../hooks/useClickableProps/useClickableProps";
 import useMergeRef from "../../hooks/useMergeRef";
 import styles from "./Label.module.scss";
 import LabelCelebrationAnimation from "./LabelCelebrationAnimation";
-import { TextType } from "../Text/TextConstants";
-
-type Sizes = "small" | "medium";
-
-const mapSizesToTextSize: Record<Sizes, TextType> = {
-  small: Text.types.TEXT3,
-  medium: Text.types.TEXT2
-};
-
-export interface LabelProps extends VibeComponentProps {
-  /**
-   * @deprecated - use className instead
-   */
-  wrapperClassName?: string;
-  /**
-   * Class name for an inner text wrapper
-   */
-  labelClassName?: string;
-  kind?: LabelKind;
-  color?: LabelColor;
-  text?: string;
-  isAnimationDisabled?: boolean;
-  isLegIncluded?: boolean;
-  onClick?: (event: React.MouseEvent<HTMLSpanElement, MouseEvent>) => void;
-  celebrationAnimation?: boolean;
-  size?: Sizes;
-}
+import { mapSizesToTextSize, LabelProps } from "./Label.types";
 
 const Label: VibeComponent<LabelProps> & {
   colors?: typeof LabelColor;
